\name{Leerkes}
\alias{Leerkes}
\docType{data}
\title{Maternal Self-Efficacy}
\description{
TIn this dataset (n = 92) the relationship between how girls were raised by there own mother and their later feelings of maternal self-efficacy, i.e. our belief in our ability to succeed in specific situations. The other variable is self-esteem (can act as mediator). All variables are scored on a continuous scale from 1 to 4. 
}
\usage{Leerkes}
\format{ 
  A data frame with 3 variables and 92 observations:
  \describe{
    \item{\code{MatCare}}{maternal care}
    \item{\code{Efficacy}}{maternal self-efficacy}
    \item{\code{Esteem}}{self-esteen}
  }
}
\references{
Leerkes, E.M. & Crockenberg, S.C. (2002). The development of maternal self-efficacy and its impact on maternal behavior. Infancy, 3, 227-247.

Howell, D.C. (2012). Statistical Methods for Psychology (8th edition). Wadsworth, Belmont, CA.
}
\examples{
summary(Leerkes)
}
\keyword{datasets}
