\name{spider}
\alias{spider}
\docType{data}
\title{Arachnophobes}
\description{24 arachnophobes were used in all. 12 were asked to play with a big hairy tarantula spider with big fangs and an evil look. Their subsequent anxiety was measured. The remaining 12 were shown only picture of the same hairy tarantula. Again, the anxiety was measured.
}
\usage{spider}
\format{
  A data frame with 2 variables and 24 observations:
  \describe{
    \item{\code{Group}}{picture vs. real spider}
    \item{\code{Anxiety}}{anxiety measure}
  }
}
\details{
Dataset from Field et al. book (p. 362).   
}
\references{
Field, A., Miles, J., & Field, Z. (2012). Discovering Statistics Using R. Sage.
}
\examples{
spider
}
\keyword{datasets}
