\name{viagra}
\alias{viagra}
\docType{data}
\title{Effects of Viagra}
\description{Participants were assigned randomly to three viagra dosages (placebo, low dosage, high dosage). The dependent variable was an objective measure of libido.}
\usage{viagra}
\format{
  A data frame with 2 variables and 15 observations:
  \describe{
    \item{\code{dose}}{viagra dosage}
    \item{\code{libido}}{objective measure of libido}
  }
}
\details{
Artificial dataset from Field et al. book (p. 401).   
}
\references{
Field, A., Miles, J, \& Field, Z. (2012). Discovering Statistics Using R. Sage.
}
\examples{
viagra
}
\keyword{datasets}
