\name{WRS2-package}
\alias{WRS2-package}
\alias{WRS2}
\docType{package}
\title{Wilcox Robust Estimation and Testing}
\description{
A user-friendly version of Wilcox' robust statistics functions (WRS package). It implements the most important robust tests as described in Rand Wilcox book "Introduction to Robust Estimation and Hypothesis Testing". 
}
\details{
\tabular{ll}{
Package: \tab WRS2\cr
Type: \tab Package\cr
Version: \tab 0.3-1\cr
Date: \tab 2014-11-29\cr
License: \tab GPL-2\cr
}
The \pkg{WRS2} package provides wrapper functions for the \pkg{WRS} package such that these functions can be applied in a user-friendly manner.
It interfaces various functions in the \pkg{WRS} package in terms of easy-to-use input arguments and corresponding S3 methods for presenting 
the output. Several robust ANOVA and ANCOVA functions have been implemented so far. Additional functions will be added subsequently. 
}
\author{
Maintainer: Patrick Mair (\email{mair@fas.harvard.edu}
}
\references{
Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.
}
\keyword{models}

