\name{diet}
\alias{diet}
\docType{data}
\title{Diet and Weight Loss}
\description{
Weight loss is studied for three different types of diets.
}
\usage{bush}
\format{
  A data frame with 7 variables and 78 observations:
  \describe{
    \item{\code{gender}}{gender}
    \item{\code{age}}{age}
    \item{\code{height}}{body height}
    \item{\code{diet.type}}{three types of diet}
    \item{\code{initial.weight}}{initial weight before diet}
    \item{\code{final.weight}}{final weight after diet}
    \item{\code{weight.loss}}{weight loss}
  }
}

\references{
Couturier, D. L., Nicholls, R., and Fernandes, M. (2018). ANOVA with R: analysis of the diet dataset. Retrieved online. 
}

\examples{
str(diet)
}
\keyword{datasets}
